Update FitReceive.py
Since we are receiving images one line at a time, and each network packet contains the line number, this version stores the line of image data in a list at an index equal to the line number.

Dropped packets mean the image is missing lines and the image is corrupt.  To solve this, we replace NULLs with neighboring elements.  This "heals" the image relatively well, especially if the number of dropped packets is small (e.g. less than 10%).